refactor(application): 지원서 목록 중복 로직을 ApplicationFormList 공유 컴포넌트로 추출
ApplicationListTab·ApplicantsListTab이 동일한 상태/핸들러/파생/렌더를 중복하고
있어 한쪽만 수정 시 갈라질 위험이 있었음. 공유 컴포넌트로 묶고 탭별 차이
(이동 대상·삭제 메시지·복제 동작·hover색)만 props로 전달. 두 탭은 얇은 래퍼로 축소.
