+++ frontend/src/pages/AdminPage/tabs/ApplicantsTab/ApplicantsListTab/ApplicantsListTab.tsx
@@ -1,240 +1,19 @@
-import React, { useEffect, useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import styled from 'styled-components';
-import expandArrow from '@/assets/images/icons/ExpandArrow.svg';
-import Plus from '@/assets/images/icons/Plus.svg';
-import Spinner from '@/components/common/Spinner/Spinner';
-import {
-  useDeleteApplication,
-  useDuplicateApplication,
-  useGetApplicationList,
-  useUpdateApplicationStatus,
-} from '@/hooks/Queries/useApplication';
-import ApplicationRowItem from '@/pages/AdminPage/components/ApplicationRow/ApplicationRowItem';
-import { ContentSection } from '@/pages/AdminPage/components/ContentSection/ContentSection';
-import * as Styled from '@/pages/AdminPage/tabs/ApplicationListTab/ApplicationListTab.styles';
-import {
-  ApplicationFormGroup,
-  ApplicationFormItem,
-  ApplicationFormStatus,
-} from '@/types/application';
+import ApplicationFormList from '@/pages/AdminPage/components/ApplicationFormList/ApplicationFormList';
 
-const MAX_INITIAL_ITEMS = 3;
-
-const ApplicationListTab = () => {
-  const { data: allforms, isLoading, isError, error } = useGetApplicationList();
-  const { mutate: deleteApplication } = useDeleteApplication();
-  const { mutate: duplicateApplication } = useDuplicateApplication();
-  const { mutate: updateStatus } = useUpdateApplicationStatus();
-
+const ApplicantsListTab = () => {
   const navigate = useNavigate();
 
-  const [isExpanded, setIsExpanded] = useState(false);
-
-  const handleGoToNewForm = () => {
-    navigate('/admin/application-list/edit');
-  };
-  const handleGoToDetailForm = (applicationFormId: string) => {
-    navigate(`/admin/applicants-list/${applicationFormId}`);
-  };
-
-  const handleDeleteApplication = (applicationFormId: string) => {
-    if (
-      window.confirm(
-        '지원서 양식을 정말 삭제하시겠습니까?\n삭제된 양식은 복구할 수 없습니다.',
-      )
-    ) {
-      deleteApplication(applicationFormId, {
-        onSuccess: () => {
-          setOpenMenuId(null);
-        },
-        onError: () => {
-          alert('삭제에 실패했습니다.');
-        },
-      });
-    }
-  };
-
-  const handleDuplicateApplication = (applicationFormId: string) => {
-    if (window.confirm('이 지원서 양식을 복제하시겠습니까?')) {
-      duplicateApplication(applicationFormId, {
-        onSuccess: () => {
-          setOpenMenuId(null);
-        },
-        onError: () => {
-          alert('지원서 복제에 실패했습니다.');
-        },
-      });
-    }
-  };
-
-  const handleToggleClick = async (
-    applicationFormId: string,
-    currentStatus: ApplicationFormStatus,
-  ) => {
-    updateStatus(
-      { applicationFormId, currentStatus },
-      {
-        onSuccess: () => {
-          setOpenMenuId(null);
-        },
-        onError: () => {
-          alert('상태 변경에 실패했습니다.');
-        },
-      },
-    );
-  };
-
-  const handleToggleExpand = () => {
-    setIsExpanded((prev) => !prev);
-  };
-
-  const [openMenuId, setOpenMenuId] = useState<string | null>(null);
-
-  const menuRef = useRef<HTMLDivElement | null>(null);
-  const handleMenuToggle = (
-    e: React.MouseEvent,
-    id: string,
-    contextPrefix: string,
-  ) => {
-    e.stopPropagation();
-    const uniqueKey = `${contextPrefix}-${id}`;
-    setOpenMenuId(openMenuId === uniqueKey ? null : uniqueKey);
-  };
-
-  useEffect(() => {
-    const handleOutsideClick = (e: MouseEvent) => {
-      if (menuRef.current && !menuRef.current.contains(e.target as Node)) {
-        setOpenMenuId(null);
-      }
-    };
-
-    if (openMenuId !== null) {
-      document.addEventListener('mousedown', handleOutsideClick);
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', handleOutsideClick);
-    };
-  }, [openMenuId]);
-
-  if (isLoading) {
-    return <Spinner />;
-  }
-  if (isError) {
-    return <div>오류가 발생했습니다: {error.message}</div>;
-  }
-
-  const formGroups: ApplicationFormGroup[] = allforms?.forms || [];
-
-  const activeForms = formGroups
-    .flatMap((group) => group.forms)
-    .filter((form) => form.status === 'ACTIVE');
-
-  const inactiveForms = formGroups
-    .flatMap((group) => group.forms)
-    .filter((form) => form.status !== 'ACTIVE');
-
-  const formsToDisplay = isExpanded
-    ? activeForms
-    : activeForms.slice(0, MAX_INITIAL_ITEMS);
-  const showExpandButton = activeForms.length > MAX_INITIAL_ITEMS;
-  const overCount = activeForms.length - MAX_INITIAL_ITEMS;
-
-  const ActiveListBody = styled(Styled.ApplicationList)`
-    border-top-left-radius: 0;
-  `;
-  const ActiveApplicationRow = styled(ApplicationRowItem)`
-    &:hover {
-      background-color: #f8f9fa;
-      &:first-child {
-        border-top-right-radius: 20px;
-      }
-    }
-  `;
-
   return (
-    <Styled.Container>
-      <div style={{ marginBottom: '24px' }}>
-        <ContentSection.Header title='지원서 목록' />
-      </div>
-      <Styled.ActiveListContainer>
-        <Styled.ActiveListTitleBox>
-          <Styled.ActiveListTitle>게시된 지원서</Styled.ActiveListTitle>
-        </Styled.ActiveListTitleBox>
-        {activeForms.length > 0 ? (
-          <ActiveListBody>
-            {formsToDisplay.map((application: ApplicationFormItem) => (
-              <ActiveApplicationRow
-                key={application.id}
-                isActive={true}
-                application={application}
-                uniqueKeyPrefix='activelist'
-                openMenuId={openMenuId}
-                menuRef={menuRef}
-                onToggleStatus={handleToggleClick}
-                onEdit={handleGoToDetailForm}
-                onMenuToggle={handleMenuToggle}
-                onDelete={handleDeleteApplication}
-                onDuplicate={handleDuplicateApplication}
-              />
-            ))}
-            {showExpandButton && (
-              <Styled.ExpandButton onClick={handleToggleExpand}>
-                {isExpanded ? '접어두기' : `펼쳐보기 (외 ${overCount}개)`}
-                <Styled.ExpandArrow
-                  src={expandArrow}
-                  $isExpanded={isExpanded}
-                />
-              </Styled.ExpandButton>
-            )}
-          </ActiveListBody>
-        ) : (
-          <ActiveListBody>
-            <Styled.MessageContainer>
-              <Styled.NoActiveFormsMessage>
-                게시된 지원서 없음
-              </Styled.NoActiveFormsMessage>
-              <Styled.SuggestionText>
-                지원서 카드 우측 메뉴에서 지원서 게시를 선택해 보세요.
-              </Styled.SuggestionText>
-            </Styled.MessageContainer>
-          </ActiveListBody>
-        )}
-      </Styled.ActiveListContainer>
-      <Styled.Header>
-        <Styled.AddButton onClick={handleGoToNewForm}>
-          새 양식 만들기 <Styled.PlusIcon src={Plus} />{' '}
-        </Styled.AddButton>
-      </Styled.Header>
-      {inactiveForms.length > 0 && (
-        <Styled.ApplicationList>
-          <Styled.ListHeader>
-            <Styled.SemesterTitle>미게시</Styled.SemesterTitle>
-            <Styled.DateHeader>
-              <Styled.Separation_Bar />
-              최종 수정 날짜
-            </Styled.DateHeader>
-          </Styled.ListHeader>
-          {inactiveForms.map((application: ApplicationFormItem) => (
-            <ApplicationRowItem
-              key={application.id}
-              application={application}
-              isActive={false}
-              uniqueKeyPrefix='inactivelist'
-              openMenuId={openMenuId}
-              menuRef={menuRef}
-              onEdit={handleGoToDetailForm}
-              onMenuToggle={handleMenuToggle}
-              onToggleStatus={handleToggleClick}
-              onDelete={handleDeleteApplication}
-              onDuplicate={handleDuplicateApplication}
-            />
-          ))}
-        </Styled.ApplicationList>
-      )}
-    </Styled.Container>
+    <ApplicationFormList
+      onEdit={(applicationFormId) =>
+        navigate(`/admin/applicants-list/${applicationFormId}`)
+      }
+      rowHoverColor='#f8f9fa'
+      deleteErrorMessage='삭제에 실패했습니다.'
+      duplicateConfirmMessage='이 지원서 양식을 복제하시겠습니까?'
+    />
   );
 };
 
-export default ApplicationListTab;
+export default ApplicantsListTab;
